requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
mistralai==1.5.0
streamlit==1.28.0
requests==2.28.2
python-dotenv==1.0.0
gunicorn==20.1.0